Statement States

+++ Projects/Application/Departments/Queries/GetDepartmentModel.cs
@@ -8,8 +8,8 @@ namespace NUCA.Projects.Application.Departments.Queries
 {
    public class GetDepartmentModel
     {
-        public int Id { get; init; }
+        public long Id { get; init; }
         public string Name { get; init; }
-        // public List<long> GroupsIds { get; init; }
+        // public List GroupsIds { get; init; }
     }
 }
