+++ Projects/Application/Users/Commands/UserModel.cs
@@ -5,6 +5,6 @@ namespace NUCA.Projects.Application.Users.Commands
     public class UserModel
     {
         public string Name { get; init; }
-        public List<int> DepartmentsIds { get; init; }
+        public List<long> DepartmentsIds { get; init; }
     }
 }
